Split PlayerView into many files
